move bytesToFelts and feltsToBytes, use it instead of duplicating
