+++ packages-ts/starknet-gauntlet/src/index.ts
@@ -11,3 +11,4 @@ export * from './provider'
 export * from './wallet'
 export * from './events'
 export * from './utils'
+export * from './encoding'
